+++ lib/id3c/cli/command/de_identify.py
@@ -9,7 +9,7 @@ from os import environ
 from sys import stdout
 from typing import Tuple
 from id3c.cli import cli
-from id3c.cli.command import (
+from id3c.cli.io.pandas import (
     load_input_from_file_or_stdin,
 )
